git commit -m "feat: improve hydration UI, fix cache zoom logic, and isolate PDAL crashes
**Hydration Tool UI (`hydrate.html`)**:
- Rewrote Leaflet interface to dynamically calculate estimates.
- Swapped center handle to a top-left (NW) corner handle for resizing.
- Added 'Total Pixels' estimate to help assess GPU texture memory loads.
- Added toggle between 'Zoom Level' and 'Custom Meters'.
- Implemented URL state serialization allowing users to bookmark regions.

**Cache Logic (`cache_manager.py`)**:
- Fixed Tier 2/Fused Zarr zoom level sorting.
- Recalculates zoom natively from Zarr array resolution rather than map width, ensuring correctly rounded integer zooms (e.g. L13 instead of L12.82).

**Provider Reliability**:
- `usgs_lidar.py`: Moved PDAL execution to a discrete `multiprocessing` worker. This isolates the main runtime from hard LibC segfaults/ArbiterErrors thrown by `pdal.Pipeline` during corrupt fetches.
- `usgs_lidar.py`: Cleaned up STAC query exception logging.
- `ncei_bag.py`: Implemented `TOPOBATHY_BAG_LRU_SIZE` (default 16) to cap concurrent LRU cache size, preventing 'Too Many Open Files' limits downstream.

**Maintenance**:
- Fixed linting/formatting errors in `usgs_3dep.py` (E501), `runtime.py` (SIM108), `vdatum.py`, and `noaa_bluetopo.py`.
- Fixed typing hints in `main.py`."

+++ src/topobathysim/api/models.py
@@ -55,7 +55,7 @@ class Tier2Detail(BaseModel):
     bytes: int
     newest: float
     oldest: float
-    by_zoom: dict[int, ZoomDetail]
+    by_zoom: dict[float, ZoomDetail]
 
 
 class ProviderDetail(BaseModel):
